Suppress warnings
Suppress the following warnings:
```
In file included from gr_common/lib/Firmata/Firmata.h:16:0,
                 from gr_common/lib/Firmata/Firmata.cpp:20:
gr_common/lib/Firmata/Boards.h:363:19: warning: extra tokens at end of #ifndef directive [enabled by default]
 #ifndef IS_PIN_SPI(p)
                   ^

gr_common/lib/Firmata/Boards.h: In function 'unsigned char writePort(byte, byte, byte)':
gr_common/lib/Firmata/Boards.h:436:1: warning: no return statement in function returning non-void [-Wreturn-type]
 }
 ^

gr_common/lib/SD/utility/SdFile.cpp: In member function 'uint8_t SdFile::rmRfStar()':
gr_common/lib/SD/utility/SdFile.cpp:910:40: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if (curPosition_ != (32*(index + 1))) {
                                        ^

gr_common/lib/Wire/Wire.cpp: In member function 'uint8_t TwoWire::requestFrom(uint8_t, uint8_t, uint8_t)':
gr_common/lib/Wire/Wire.cpp:182:20: warning: unused variable 'b' [-Wunused-variable]
               char b = RIIC0.ICDRR; // dummy read
                    ^

wrbb_mruby/sWiFi.cpp:2544:2: warning: this decimal constant is unsigned only in ISO C90 [enabled by default]
  time -= 2208988800; // conversion to Unixtime
  ^

wrbb_mruby/sWiFi.cpp: In function 'mrb_wifi_server(mrb_state*, mrb_value)':
wrbb_mruby/sWiFi.cpp:2212:28: warning: 'c' may be used uninitialized in this function [-Wmaybe-uninitialized]
        WiFiData[wifCnt] = c;
                            ^

gr_common/lib/Wire/utility/twi_rx.c: In function 'twi_rx_read':
gr_common/lib/Wire/utility/twi_rx.c:164:18: warning: unused variable 'temp' [-Wunused-variable]
             char temp = RIIC0.ICDRR;
                  ^
gr_common/lib/Wire/utility/twi_rx.c: In function 'twi_rx_write':
gr_common/lib/Wire/utility/twi_rx.c:268:22: warning: unused variable 'temp' [-Wunused-variable]
                 char temp = RIIC0.ICDRR; // dummy read
                      ^
gr_common/lib/Wire/utility/twi_rx.c: In function 'INT_Excep_RIIC0_RXI0':
gr_common/lib/Wire/utility/twi_rx.c:347:14: warning: unused variable 'b' [-Wunused-variable]
         char b = RIIC0.ICDRR; // dummy read
              ^

WavMp3p/libmad-0.15.1b/frame.c: In function 'mad_header_decode':
WavMp3p/libmad-0.15.1b/frame.c:330:3: warning: implicit declaration of function 'copy_next_header' [-Wimplicit-function-declaration]
   copy_next_header();
   ^
WavMp3p/libmad-0.15.1b/frame.c:331:3: warning: implicit declaration of function 'get_next_frame' [-Wimplicit-function-declaration]
   ptr = stream->next_frame = get_next_frame();
   ^
WavMp3p/libmad-0.15.1b/frame.c:331:28: warning: assignment makes pointer from integer without a cast [enabled by default]
   ptr = stream->next_frame = get_next_frame();
                            ^
WavMp3p/libmad-0.15.1b/frame.c:442:3: warning: implicit declaration of function 'read_frame' [-Wimplicit-function-declaration]
   read_frame(N);
   ^

WavMp3p/utility/wavmp3p_play_mp3.c: In function 'read_frame':
WavMp3p/utility/wavmp3p_play_mp3.c:144:2: warning: dereferencing type-punned pointer will break strict-aliasing rules [-Wstrict-aliasing]
  *(int*)mp3_work = 0;
  ^
WavMp3p/utility/wavmp3p_play_wav.c: In function 'wavmp3p_play_wav':
WavMp3p/utility/wavmp3p_play_wav.c:66:2: warning: dereferencing type-punned pointer will break strict-aliasing rules [-Wstrict-aliasing]
  unsigned long n = *(unsigned long *)read_buffer;
  ^
WavMp3p/utility/wavmp3p_play_wav.c:99:2: warning: dereferencing type-punned pointer will break strict-aliasing rules [-Wstrict-aliasing]
  chunk_size = *(unsigned long *)read_buffer;
  ^
```

+++ firmware_release/gr_common/lib/SD/utility/SdFile.cpp
@@ -883,7 +883,7 @@ uint8_t SdFile::rmRfStar(void) {
     SdFile f;
 
     // remember position
-    uint16_t index = curPosition_/32;
+    uint32_t index = curPosition_/32;
 
     dir_t* p = readDirCache();
     if (!p) return false;
@@ -897,7 +897,7 @@ uint8_t SdFile::rmRfStar(void) {
     // skip if part of long file name or volume label in root
     if (!DIR_IS_FILE_OR_SUBDIR(p)) continue;
 
-    if (!f.open(this, index, O_READ)) return false;
+    if (!f.open(this, (uint16_t)index, O_READ)) return false;
     if (f.isSubDir()) {
       // recursively delete
       if (!f.rmRfStar()) return false;
